requirements.txt
streamlit
numpy
pandas
torch
torchvision
Pillow
openpyxl
matplotlib
scikit-learn